collectors/requirements.txt
pandas>=2.0.0
google-cloud-bigquery>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
